Time: 141 ms (43.92%), Space: 35.3 MB (15.28%) - LeetHub

+++ 1509-minimum-difference-between-largest-and-smallest-value-in-three-moves/1509-minimum-difference-between-largest-and-smallest-value-in-three-moves.cpp
@@ -5,9 +5,9 @@ public:
         if(n<=4) return 0;
         sort(nums.begin(),nums.end());
         
-        long long ans = min(nums[n-1]-nums[3],min(nums[n-4]-nums[0],min(nums[n-2]-nums[2],nums[n-3]-nums[1])));
+        int ans = min(nums[n-1]-nums[3],min(nums[n-4]-nums[0],min(nums[n-2]-nums[2],nums[n-3]-nums[1])));
         
-        return (int)ans;
+        return ans;
         
     }
 };
